requirements.txt
streamlit
Pillow==8.1.1
tensorflow-cpu
numpy
pytest-shutil
